requirements.txt
torch --pre --index-url https://download.pytorch.org/whl/nightly/cu121
torchvision
torchaudio
transformers
datasets
pandas
scikit-learn
matplotlib
seaborn
tqdm
pynvml